libc: gen/bzero.c; ANSI, size_t length. Update manpage.
Update the function prototype for bzero() to:

void bzero(void *b, size_t length);

The length parameter is updated to a size_t.

De-register and ANSI function cleanup.

Update bzero(3) function prototype in bstring(3) manual page.
Small typo and whitespace fixes as well. Bump the modified date.

Added back the missing 2.11BSD SCCS version tag to bzero.c.

+++ include/strings.h
@@ -40,7 +40,7 @@ typedef	unsigned int	size_t;
 
 int	 bcmp(const void *, const void *, size_t);
 void	 bcopy(const void *, void *, size_t);
-void	 bzero(void *, unsigned long);
+void	 bzero(void *, size_t);
 int	 ffs(int);
 char	*index(const char *, int);
 char	*rindex(const char *, int);
